+++ FluidWarfare.Editor.Windows/Panels/Inspector/InspectorPanel.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace FluidWarfare.Editor.Windows.Panels.Inspector;
+
+public sealed partial class InspectorPanel : UserControl
+{
+    public InspectorPanel()
+    {
+        InitializeComponent();
+    }
+}