+++ FluidWarfare.Editor.Windows/Panels/Project/ProjectPanel.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace FluidWarfare.Editor.Windows.Panels.Project;
+
+public sealed partial class ProjectPanel : UserControl
+{
+    public ProjectPanel()
+    {
+        InitializeComponent();
+    }
+}