+++ FluidWarfare.Editor.Windows/Panels/Viewport/ViewportPlaceholderPanel.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace FluidWarfare.Editor.Windows.Panels.Viewport;
+
+public sealed partial class ViewportPlaceholderPanel : UserControl
+{
+    public ViewportPlaceholderPanel()
+    {
+        InitializeComponent();
+    }
+}